refactor: extract shared lib, split test cases, add CLI flags
- Extract shared utilities (constants, read, report, arg parsing) to tests/lib.py
- Split 30 e2e test cases into tests/cases.py, grouped by area (STK, FMT, ITV, DPL)
- Add --area=STK flag to filter tests by area prefix
- Add --fail-fast flag to stop on first failure
- Add elapsed time to summary output
- Add progress indicator ([1/30] running...) for live mode
- Both runners now use shared report writer with total elapsed time

run_e2e.py reduced from 967 to 193 lines. All 7 smoke checks and
30 e2e tests pass.

Closes #10
