+++ tests/run_e2e.py
@@ -9,8 +9,10 @@ Usage:
   py tests/run_e2e.py                    # run all non-skipped tests
   py tests/run_e2e.py STK-01             # run one test by short ID
   py tests/run_e2e.py STK-01 FMT-01      # run multiple
+  py tests/run_e2e.py --area=STK          # run all stack tests
   py tests/run_e2e.py --dry-run          # print prompt, skip claude call
   py tests/run_e2e.py --offline          # validate test infrastructure without API
+  py tests/run_e2e.py --fail-fast        # stop on first failure
 
 Offline mode validates:
   - All referenced template files exist and are non-empty
@@ -23,18 +25,13 @@ each sending ~10–20K tokens to Claude).
 """
 
 import datetime
-import io
 import os
 import subprocess
 import sys
 import time
 
-ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-
-
-def read(rel_path):
-    with io.open(os.path.join(ROOT, rel_path), encoding="utf-8") as f:
-        return f.read()
+from lib import ROOT, PASS, FAIL, SKIP, ERR, REPORT_TRUNCATION, read, parse_args
+from cases import ALL_TESTS
 
 
 def build_prompt(stack_file, answers, output_file="formats/agents.md",
@@ -62,8 +59,6 @@ def build_prompt(stack_file, answers, output_file="formats/agents.md",
 
 
 def run_claude(prompt, timeout=180):
-    # Pass the prompt via stdin to avoid shell quoting issues on Windows.
-    # `claude -p` reads stdin when no positional prompt argument is given.
     result = subprocess.run(
         "claude -p --no-session-persistence",
         input=prompt,
@@ -88,743 +83,10 @@ def check_assertions(output, required=(), forbidden=()):
     return failures
 
 
-# ---------------------------------------------------------------------------
-# Test cases
-#
-# Each test entry has these fields:
-#   id          short ID used on the command line (e.g. STK-01)
-#   spec        the spec file this test implements
-#   stack       stack template file fed to the agent
-#   answers     interview answers the agent receives
-#   output_file output format template (defaults to formats/CLAUDE.md)
-#   extra_files additional template files to include in the prompt
-#   required    strings that MUST appear in the output — the test fails if any
-#               are missing; used to assert that key sections and rules are
-#               present
-#   forbidden   strings that MUST NOT appear in the output — the test fails if
-#               any are found; used to assert that:
-#               - unfilled template placeholders were replaced by real values
-#               - rules from a different output format did not leak in
-#               - sections that do not apply to this stack are absent
-#   skip        if present, the test is skipped with this message
-# ---------------------------------------------------------------------------
-
-TESTS = [
-    # -------------------------------------------------------------------------
-    # FastAPI
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-01",
-        "spec": "SAIT-E2E-STK-01-001A",
-        "stack": "stack/python-fastapi.md",
-        "answers": {
-            "Project name": "OrderService",
-            "Owner": "Platform team",
-            "Repo": "github.com/acme/order-service",
-            "Deployment": "Docker to Kubernetes (cloud)",
-            "Database": "PostgreSQL via SQLAlchemy 2 + Alembic",
-            "Auth": "JWT bearer tokens",
-            "Feature flags": "no",
-            "Messaging": "no",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Python 3.11", "FastAPI", "SQLAlchemy",
-            "pytest",
-            "async def",
-            "pydantic-settings",
-            "feat/",
-        ],
-        # these are raw template placeholders — if any appear, the agent
-        # failed to substitute the interview answers into the output
-        "forbidden": ["[your project]", "[owner]", "[repo]", "[platform]"],
-    },
-    # -------------------------------------------------------------------------
-    # Go Echo
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-02",
-        "spec": "SAIT-E2E-STK-02-001A",
-        "stack": "stack/go-echo.md",
-        "answers": {
-            "Project name": "MetricsHub",
-            "Owner": "Infrastructure team",
-            "Repo": "github.com/acme/metricshub",
-            "Deployment": "Docker to Kubernetes (cloud)",
-            "Database": "PostgreSQL via sqlc + pgx",
-            "Auth": "JWT bearer tokens",
-            "Feature flags": "yes — OpenFeature Go SDK",
-            "Messaging": "no",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Go 1.22", "Echo v4", "sqlc", "pgx",
-            "cmd/", "internal/",
-            "errgroup", "SIGTERM",
-            "OpenFeature",
-            "go.sum",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Django
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-03",
-        "spec": "SAIT-E2E-STK-03-001A",
-        "stack": "stack/python-django.md",
-        "answers": {
-            "Project name": "CatalogService",
-            "Framework": "Django 5.x + Django REST Framework",
-            "Database": "PostgreSQL via Django ORM",
-            "Auth": "JWT via djangorestframework-simplejwt",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Django", "REST framework",
-            "select_related", "prefetch_related",
-            "migration",
-            "simplejwt",
-            "pytest-django", "django_db",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Node Express
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-04",
-        "spec": "SAIT-E2E-STK-04-001A",
-        "stack": "stack/node-express.md",
-        "answers": {
-            "Project name": "NotificationService",
-            "Language": "TypeScript",
-            "Database": "PostgreSQL",
-            "Auth": "JWT",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Node", "Express", "TypeScript", "Zod",
-            "middleware",
-            "supertest",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # React SPA
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-05",
-        "spec": "SAIT-E2E-STK-05-001A",
-        "stack": "stack/spa-react.md",
-        "answers": {
-            "Project name": "DashboardApp",
-            "Language": "TypeScript",
-            "State management": "Zustand",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "React", "TypeScript", "Zustand",
-            "component",
-            "React Testing Library",
-            "accessibility",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Next.js
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-06",
-        "spec": "SAIT-E2E-STK-06-001A",
-        "stack": "stack/full-nextjs.md",
-        "answers": {
-            "Project name": "StorefrontApp",
-            "Language": "TypeScript",
-            "Rendering strategy": "App Router with Server Components",
-            "Database": "Prisma + PostgreSQL",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Next.js", "React", "TypeScript", "Prisma",
-            "App Router", "Server Component",
-            "Route Handler",
-            "feat/",
-        ],
-        # Next.js App Router project — pages/ is the old Pages Router convention
-        # and must not appear; its presence means the model used the wrong routing model
-        "forbidden": ["pages/"],
-    },
-    # -------------------------------------------------------------------------
-    # Astro
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-07",
-        "spec": "SAIT-E2E-STK-07-001A",
-        "stack": "stack/static-site-astro.md",
-        "answers": {
-            "Project name": "TechBlog",
-            "Language": "TypeScript",
-            "Content": "Markdown + MDX",
-            "Integrations": "React islands for interactive components",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions",
-            "Astro", "TypeScript", "MDX",
-            "src/content",
-            "client:",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Go gRPC
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-08",
-        "spec": "SAIT-E2E-STK-08-001A",
-        "stack": "stack/go-grpc.md",
-        "answers": {
-            "Project name": "PaymentGateway",
-            "Language": "Go",
-            "Communication": "gRPC (internal service-to-service)",
-            "Auth": "mTLS",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Go", "gRPC", "proto",
-            "interceptor",
-            "status code",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Flutter
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-09",
-        "spec": "SAIT-E2E-STK-09-001A",
-        "stack": "stack/mobile-flutter.md",
-        "answers": {
-            "Project name": "FieldSurveyApp",
-            "Language": "Dart",
-            "State management": "Riverpod",
-            "Platforms": "iOS and Android",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Flutter", "Dart", "Riverpod",
-            "widget",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Go library
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-10",
-        "spec": "SAIT-E2E-STK-10-001A",
-        "stack": "stack/go-lib.md",
-        "answers": {
-            "Project name": "retrykit",
-            "Language": "Go",
-            "Distribution": "Go module (pkg.go.dev)",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Go",
-            "doc comment",
-            "vX.Y.Z",
-            "go.sum",
-            "feat/",
-        ],
-        # go-lib is a library, not a deployable service — Dockerfile and a
-        # Deployment section have no place here
-        "forbidden": ["Dockerfile", "## Deployment"],
-    },
-    # -------------------------------------------------------------------------
-    # Flask
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-11",
-        "spec": "SAIT-E2E-STK-11-001A",
-        "stack": "stack/python-flask.md",
-        "answers": {
-            "Project name": "ReportingAPI",
-            "Language": "Python",
-            "Database": "PostgreSQL via Flask-SQLAlchemy",
-            "Auth": "JWT",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Flask", "Python", "SQLAlchemy",
-            "create_app",
-            "blueprint",
-            "pytest-flask",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Generic Python service
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-12",
-        "spec": "SAIT-E2E-STK-12-001A",
-        "stack": "stack/python-service.md",
-        "answers": {
-            "Project name": "DataPipelineWorker",
-            "Language": "Python",
-            "Framework": "none (plain Python service)",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Python",
-            "ruff",
-            "pytest",
-            "feat/",
-        ],
-        # python-service is framework-agnostic — if a specific framework name
-        # appears, the model pulled rules from the wrong template
-        "forbidden": ["FastAPI", "Flask", "Django"],
-    },
-    # -------------------------------------------------------------------------
-    # Python gRPC
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-13",
-        "spec": "SAIT-E2E-STK-13-001A",
-        "stack": "stack/python-grpc.md",
-        "answers": {
-            "Project name": "MLInferenceService",
-            "Language": "Python",
-            "Communication": "gRPC (internal service-to-service)",
-            "Auth": "mTLS",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Python", "gRPC", "proto",
-            "grpcio",
-            "servicer",
-            "interceptor",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Celery worker
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-14",
-        "spec": "SAIT-E2E-STK-14-001A",
-        "stack": "stack/python-celery-worker.md",
-        "answers": {
-            "Project name": "EmailDispatchWorker",
-            "Language": "Python",
-            "Broker": "Redis",
-            "Result backend": "Redis",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Python", "Celery", "Redis",
-            "idempoten",
-            "retry", "max_retries",
-            "Celery Beat",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Python library
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-15",
-        "spec": "SAIT-E2E-STK-15-001A",
-        "stack": "stack/python-lib.md",
-        "answers": {
-            "Project name": "validify",
-            "Language": "Python",
-            "Distribution": "PyPI package",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Python",
-            "pyproject.toml",
-            "docstring",
-            "feat/",
-        ],
-        # python-lib is a library, not a deployable service — Dockerfile and a
-        # Deployment section have no place here
-        "forbidden": ["Dockerfile", "## Deployment"],
-    },
-    # -------------------------------------------------------------------------
-    # Generic Go service (stdlib)
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-16",
-        "spec": "SAIT-E2E-STK-16-001A",
-        "stack": "stack/go-service.md",
-        "answers": {
-            "Project name": "HealthCheckService",
-            "Language": "Go",
-            "HTTP framework": "none (standard library only)",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Go",
-            "net/http",
-            "graceful shutdown",
-            "go.sum",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Hugo
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-17",
-        "spec": "SAIT-E2E-STK-17-001A",
-        "stack": "stack/static-site-hugo.md",
-        "answers": {
-            "Project name": "DocumentationSite",
-            "Language": "Markdown + Go templates",
-            "Theme": "custom",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Git conventions", "## Code conventions",
-            "Hugo",
-            "archetype",
-            "shortcode",
-            "feat/",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # Node.js library
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-18",
-        "spec": "SAIT-E2E-STK-18-001A",
-        "stack": "stack/nodejs-lib.md",
-        "answers": {
-            "Project name": "parsekit",
-            "Language": "TypeScript",
-            "Distribution": "npm package",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Node", "TypeScript",
-            "exports",
-            "vitest",
-            "feat/",
-        ],
-        # node-lib is a library, not a deployable service — Dockerfile and a
-        # Deployment section have no place here
-        "forbidden": ["Dockerfile", "## Deployment"],
-    },
-    # -------------------------------------------------------------------------
-    # Rust library
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-19",
-        "spec": "SAIT-E2E-STK-19-001A",
-        "stack": "stack/rust-lib.md",
-        "answers": {
-            "Project name": "byteparser",
-            "Language": "Rust",
-            "Distribution": "crates.io",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions", "## Testing",
-            "Rust",
-            "pub",
-            "doc comment",
-            "cargo test", "cargo clippy",
-            "feat/",
-        ],
-        # rust-lib is a library — web frameworks are service concerns and must
-        # not appear here; Tokio is excluded because the model may reference it
-        # in a negative context ("do not add Tokio unless async is needed")
-        "forbidden": ["axum", "actix"],
-    },
-    # -------------------------------------------------------------------------
-    # HTMX
-    # -------------------------------------------------------------------------
-    {
-        "id": "STK-20",
-        "spec": "SAIT-E2E-STK-20-001A",
-        "stack": "stack/htmx.md",
-        "answers": {
-            "Project name": "AdminDashboard",
-            "Backend language": "Python (Flask)",
-            "Templating engine": "Jinja2",
-            "Client-side state": "Alpine.js",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands",
-            "## Git conventions", "## Code conventions",
-            "HTMX", "Jinja2",
-            "hx-",
-            "fragment",
-            "Alpine",
-            "feat/",
-        ],
-        # HTMX is a server-rendered hypermedia approach — JSON APIs and SPA
-        # routing patterns belong to a different architecture and must not appear
-        "forbidden": ["JSON API", "SPA routing"],
-    },
-    # -------------------------------------------------------------------------
-    # FMT-01  CLAUDE.md (Claude Code)
-    # -------------------------------------------------------------------------
-    {
-        "id": "FMT-01",
-        "spec": "SAIT-E2E-FMT-01-001A",
-        "stack": "stack/python-fastapi.md",
-        "output_file": "formats/agents.md",
-        "answers": {
-            "Project name": "OrderService",
-            "Database": "PostgreSQL via SQLAlchemy 2",
-            "Auth": "JWT bearer tokens",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Architecture", "## Commands",
-            "## Code conventions", "## Git conventions", "## Testing",
-            "Python", "FastAPI",
-            "feat/",
-        ],
-        # CLAUDE.md has no frontmatter or tool-specific directives — these
-        # keywords belong to Cursor (.mdc) format and must not appear here
-        "forbidden": ["alwaysApply", "applyTo", "globs:"],
-    },
-    # -------------------------------------------------------------------------
-    # FMT-02  AGENTS.md (Codex CLI)
-    # -------------------------------------------------------------------------
-    {
-        "id": "FMT-02",
-        "spec": "SAIT-E2E-FMT-02-001A",
-        "stack": "stack/python-fastapi.md",
-        "output_file": "formats/agents.md",
-        "answers": {
-            "Project name": "OrderService",
-            "Database": "PostgreSQL via SQLAlchemy 2",
-            "Auth": "JWT bearer tokens",
-            "Output format": "AGENTS.md",
-        },
-        "required": [
-            "## Stack", "## Commands", "## Code conventions",
-            "## Git conventions", "## Testing",
-            "Python", "FastAPI",
-            "feat/",
-        ],
-        # AGENTS.md has no frontmatter or Cursor-specific directives — these
-        # keywords would indicate the model applied the wrong output format
-        "forbidden": ["alwaysApply", "applyTo", "frontmatter"],
-    },
-    # FMT-03, FMT-04, FMT-05 removed — cursorrules, copilot, generic
-    # formats dropped in favour of CLAUDE.md + AGENTS.md only
-    {
-        "id": "FMT-03",
-        "spec": "SAIT-E2E-FMT-03-001A",
-        "skip": True,
-    },
-    {
-        "id": "FMT-04",
-        "spec": "SAIT-E2E-FMT-04-001A",
-        "skip": True,
-    },
-    {
-        "id": "FMT-05",
-        "spec": "SAIT-E2E-FMT-05-001A",
-        "skip": True,
-    },
-    # -------------------------------------------------------------------------
-    # ITV-02  DEFAULTED sections pre-filled without user input
-    # Provide only REQUIRED project identity answers — no git convention
-    # answers. Assert that git conventions appear in output anyway (pre-filled
-    # from base/git.md via the stack dependency chain).
-    # -------------------------------------------------------------------------
-    {
-        "id": "ITV-02",
-        "spec": "SAIT-INT-ITV-02-001A",
-        "stack": "stack/python-fastapi.md",
-        "answers": {
-            "Project name": "InventoryService",
-            "Owner": "Backend team",
-            "Deployment": "Docker",
-            "Database": "PostgreSQL via SQLAlchemy 2",
-            "Auth": "JWT bearer tokens",
-            "Output format": "CLAUDE.md",
-            # Deliberately omit git convention answers — should be pre-filled
-        },
-        "required": [
-            "## Git conventions",
-            "feat/",         # from base/git.md DEFAULTED section
-            "## Stack", "## Commands", "## Code conventions", "## Testing",
-            "Python", "FastAPI",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # ITV-03  Interview answer overrides stack template rule
-    # Provide a contradicting answer for the test runner. Assert the override
-    # wins. Note: pytest may still appear in ## Commands — that is acceptable;
-    # the Testing *section* must use unittest.
-    # -------------------------------------------------------------------------
-    {
-        "id": "ITV-03",
-        "spec": "SAIT-INT-ITV-03-001A",
-        "stack": "stack/python-fastapi.md",
-        "answers": {
-            "Project name": "InventoryService",
-            "Owner": "Backend team",
-            "Deployment": "Docker",
-            "Database": "PostgreSQL via SQLAlchemy 2",
-            "Auth": "JWT bearer tokens",
-            "Test runner override": "Use unittest instead of pytest for all tests",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Stack", "## Commands", "## Git conventions",
-            "## Code conventions", "## Testing",
-            "unittest",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # DPL-01  Cloud deployment target
-    # -------------------------------------------------------------------------
-    {
-        "id": "DPL-01",
-        "spec": "SAIT-E2E-DPL-01-001A",
-        "stack": "stack/node-express.md",
-        "extra_files": ["base/deployment.md"],
-        "answers": {
-            "Project name": "PublicAPIService",
-            "Owner": "Platform team",
-            "Language": "TypeScript",
-            "Deployment target": "cloud",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Deployment",
-            "cloud",
-        ],
-        # cloud target means public infrastructure only — private CA and
-        # air-gap/offline assumptions must not appear in the output
-        "forbidden": ["private CA", "air-gap", "offline"],
-    },
-    # -------------------------------------------------------------------------
-    # DPL-02  Hybrid deployment target
-    # -------------------------------------------------------------------------
-    {
-        "id": "DPL-02",
-        "spec": "SAIT-E2E-DPL-02-001A",
-        "stack": "stack/java-spring-boot.md",
-        "extra_files": ["base/deployment.md"],
-        "answers": {
-            "Project name": "InternalPlatformAPI",
-            "Owner": "Platform team",
-            "Language": "Java",
-            "Deployment target": "hybrid (on-premises + cloud)",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            "## Deployment",
-            "hybrid",
-        ],
-        "forbidden": [],
-    },
-    # -------------------------------------------------------------------------
-    # DPL-03  Offline / air-gapped deployment target
-    # -------------------------------------------------------------------------
-    {
-        "id": "DPL-03",
-        "spec": "SAIT-E2E-DPL-03-001A",
-        "stack": "stack/python-fastapi.md",
-        "extra_files": ["base/deployment.md"],
-        "answers": {
-            "Project name": "SecureIngestService",
-            "Owner": "Platform team",
-            "Language": "Python",
-            "Deployment target": "offline (air-gapped, no internet access)",
-            "Output format": "CLAUDE.md",
-        },
-        "required": [
-            # model may embed deployment info in project identity rather than
-            # a dedicated ## Deployment section — check for the target keyword
-            "offline",
-            "air-gap",
-        ],
-        # offline/air-gapped target means no public internet — public
-        # registries must not appear; ACM removed because model may reference
-        # it negatively ("do not use ACM in air-gapped environments")
-        "forbidden": ["Let's Encrypt", "Docker Hub"],
-    },
-]
-
-# ---------------------------------------------------------------------------
-# Runner
-# ---------------------------------------------------------------------------
-
-PASS = "PASS"
-FAIL = "FAIL"
-SKIP = "SKIP"
-ERR  = "ERR "
-
-
 def validate_test_offline(test):
     """Validate test infrastructure without calling Claude."""
     failures = []
-    tid = test["id"]
 
-    # Required fields
     for field in ("id", "spec", "stack", "answers", "required"):
         if field not in test:
             failures.append(f"  missing field: {field!r}")
@@ -832,7 +94,6 @@ def validate_test_offline(test):
     if failures:
         return FAIL, "\n".join(failures)
 
-    # Template files exist and are non-empty
     stack_file = os.path.join(ROOT, test["stack"])
     if not os.path.isfile(stack_file):
         failures.append(f"  stack file missing: {test['stack']}")
@@ -849,11 +110,9 @@ def validate_test_offline(test):
         if not os.path.isfile(ef_path):
             failures.append(f"  extra file missing: {ef}")
 
-    # Assertions are non-empty
     if not test.get("required"):
         failures.append("  required list is empty")
 
-    # Prompt builds without error
     if not failures:
         try:
             prompt = build_prompt(
@@ -912,94 +171,82 @@ def run_test(test, dry_run=False, offline=False):
     return PASS, f"{elapsed:.1f}s", elapsed, None
 
 
+def render_fail(r):
+    lines = []
+    elapsed_str = f"  ({r['elapsed']:.1f}s)" if r["elapsed"] else ""
+    lines.append(f"### {r['status']}  {r['id']}{elapsed_str}")
+    lines.append("")
+    lines.append("**Expected**:")
+    lines.append("")
+    lines.append("```")
+    for line in r["detail"].splitlines():
+        lines.append(line)
+    lines.append("```")
+    lines.append("")
+    lines.append(f"**Observed** (first {REPORT_TRUNCATION} chars of model output):")
+    lines.append("")
+    lines.append("```")
+    observed = (r["output"] or "")[:REPORT_TRUNCATION].replace("```", "~~~")
+    lines.append(observed)
+    lines.append("```")
+    lines.append("")
+    return lines
+
+
+def render_err(r):
+    return [f"### {r['status']}  {r['id']}", "", f"**Error**: {r['detail']}", ""]
+
+
+def render_skip(r):
+    return [f"### {r['status']}  {r['id']}", "", f"**Skipped**: {r['detail']}", ""]
+
+
+def render_pass(r):
+    elapsed_str = f"  ({r['detail']})" if r["detail"] else ""
+    return [f"### {r['status']}  {r['id']}{elapsed_str}", ""]
+
+
 def write_report(run_results, started_at, dry_run):
-    reports_dir = os.path.join(ROOT, "tests", "reports")
-    os.makedirs(reports_dir, exist_ok=True)
-
-    ts = started_at.strftime("%Y-%m-%dT%H-%M-%S")
-    report_path = os.path.join(reports_dir, f"{ts}-e2e.md")
-
-    passed  = sum(1 for r in run_results if r["status"] == PASS)
-    failed  = sum(1 for r in run_results if r["status"] == FAIL)
-    skipped = sum(1 for r in run_results if r["status"] == SKIP)
-    errored = sum(1 for r in run_results if r["status"] == ERR)
-    total   = len(run_results)
-
-    lines = [
-        "# E2E Test Report",
-        "",
-        f"**Date**: {started_at.strftime('%Y-%m-%d %H:%M:%S')}  ",
-        f"**Runner**: run_e2e.py  ",
-        f"**Mode**: {'dry-run' if dry_run else 'live'}  ",
-        f"**Tests run**: {total}",
-        "",
-        "## Summary",
-        "",
-        (f"{total} tests — {passed} passed  {failed} failed  "
-         f"{skipped} skipped  {errored} errors"),
-        "",
-        "---",
-        "",
-        "## Results",
-        "",
-    ]
-
-    for r in run_results:
-        elapsed_str = f"  ({r['elapsed']:.1f}s)" if r["elapsed"] else ""
-        lines.append(f"### {r['status']}  {r['id']}{elapsed_str}")
-        lines.append("")
-        if r["status"] == FAIL:
-            lines.append("**Expected**:")
-            lines.append("")
-            lines.append("```")
-            for line in r["detail"].splitlines():
-                lines.append(line)
-            lines.append("```")
-            lines.append("")
-            lines.append("**Observed** (first 1000 chars of model output):")
-            lines.append("")
-            lines.append("```")
-            # replace triple backticks in model output so they don't close
-            # the report's own code fence
-            observed = (r["output"] or "")[:1000].replace("```", "~~~")
-            lines.append(observed)
-            lines.append("```")
-            lines.append("")
-        elif r["status"] == ERR:
-            lines.append(f"**Error**: {r['detail']}")
-            lines.append("")
-        elif r["status"] == SKIP:
-            lines.append(f"**Skipped**: {r['detail']}")
-            lines.append("")
-
-    with io.open(report_path, "w", encoding="utf-8") as f:
-        f.write("\n".join(lines))
-
-    print(f"\nReport: {os.path.relpath(report_path, ROOT)}")
+    from lib import write_report as _write_report
+    _write_report(run_results, started_at, "e2e", {
+        PASS: render_pass,
+        FAIL: render_fail,
+        SKIP: render_skip,
+        ERR: render_err,
+    })
 
 
 def main():
     started_at = datetime.datetime.now()
 
-    args = sys.argv[1:]
-    dry_run = "--dry-run" in args
-    offline = "--offline" in args
-    filter_ids = [a for a in args if not a.startswith("--")]
+    filter_ids, flags, area = parse_args(sys.argv[1:])
+    dry_run = "dry-run" in flags
+    offline = "offline" in flags
+    fail_fast = "fail-fast" in flags
 
-    tests = TESTS
+    tests = ALL_TESTS
     if filter_ids:
-        tests = [t for t in TESTS if t["id"] in filter_ids]
+        tests = [t for t in ALL_TESTS if t["id"] in filter_ids]
         if not tests:
             print(f"No tests matched: {filter_ids}")
+            sys.exit(1)
+    elif area:
+        tests = [t for t in ALL_TESTS if t["id"].startswith(area)]
+        if not tests:
+            print(f"No tests matched area: {area}")
             sys.exit(1)
 
     results = {PASS: 0, FAIL: 0, SKIP: 0, ERR: 0}
     run_results = []
 
-    print(f"Running {len(tests)} test(s)...\n")
+    total = len(tests)
+    print(f"Running {total} test(s)...\n")
 
-    for test in tests:
+    for i, test in enumerate(tests, 1):
         tid = test["id"]
+        if not dry_run and not offline:
+            print(f"  [{i}/{total}] {tid} running...", end="\r", flush=True)
+
         status, detail, elapsed, output = run_test(test, dry_run=dry_run, offline=offline)
         results[status] += 1
         run_results.append({
@@ -1015,13 +262,19 @@ def main():
             print(f"  {status}  {tid}")
             print(detail)
 
-    total = sum(results.values())
+        if fail_fast and status in (FAIL, ERR):
+            print("\n  --fail-fast: stopping after first failure")
+            break
+
+    elapsed = (datetime.datetime.now() - started_at).total_seconds()
+    total_run = sum(results.values())
     print(
-        f"\n{total} tests — "
+        f"\n{total_run} tests — "
         f"{results[PASS]} passed  "
         f"{results[FAIL]} failed  "
         f"{results[SKIP]} skipped  "
         f"{results[ERR]} errors"
+        f"  ({elapsed:.1f}s)"
     )
 
     write_report(run_results, started_at, dry_run or offline)
